backend/requirements.txt
Django==3.0.6
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
PyJWT==v1.7.1
cryptography
matplotlib==3.2.1
pandas==1.0.3
psycopg2-binary==2.8.5
pulp==2.1
numpy==1.18.4
whitenoise
